+++ CHANGES.rst
@@ -33,6 +33,8 @@ Bug fixes:
 
 - No longer use deprecated import for getSite/setSite.
   [jensens]
+- Python 2 / 3 compatible imports.
+  [rudaporto]
 
 - Update code to follow Plone styleguide.
   [gforcada]